# server/internal/platform/httpx/openapi.yaml
openapi: 3.1.0
info:
  title: Ship API
  version: 0.1.0
  description: The contract exposed by the Ship control-plane API.
servers:
  - url: /
paths:
...
  /users:
    get:
      operationId: listUsers
      x-ship-permission: users.read
      security: [{ sessionCookie: [] }]
      summary: List local accounts
      responses:
        "200":
          description: Local accounts.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserList" }
        "403":
          description: Permission denied.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    post:
      operationId: createUser
      x-ship-permission: users.manage
      security: [{ sessionCookie: [] }]
      summary: Create a local account
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateUserRequest" }
      responses:
        "201":
          description: Account created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "403":
          description: Permission denied.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: Email already exists.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /users/{id}/role:
    patch:
      operationId: changeUserRole
      x-ship-permission: users.manage
      security: [{ sessionCookie: [] }]
      summary: Change a local account role
      parameters:
        - $ref: "#/components/parameters/UserID"
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChangeRoleRequest" }
      responses:
        "200":
          description: Role changed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "403":
          description: Permission denied.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: User not found.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /users/{id}:
    delete:
      operationId: disableUser
      x-ship-permission: users.manage
      security: [{ sessionCookie: [] }]
      summary: Disable a local account and invalidate its sessions
      parameters:
        - $ref: "#/components/parameters/UserID"
        - $ref: "#/components/parameters/CSRFToken"
      responses:
        "204": { description: Account disabled. }
        "403":
          description: Permission denied.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: User not found.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: The last active owner cannot be disabled.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-ship_session
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string, minLength: 32 }
    UserID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  schemas:
...
    Role:
      type: string
      enum: [owner, admin, deployer, viewer]
    User:
      type: object
      required: [id, email, role]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        role: { $ref: "#/components/schemas/Role" }
    ManagedUser:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          required: [disabled, createdAt]
          properties:
            disabled: { type: boolean }
            createdAt: { type: string, format: date-time }
...
    CreateUserRequest:
      type: object
      required: [email, password, role]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 12, maxLength: 128 }
        role: { $ref: "#/components/schemas/Role" }
    ChangeRoleRequest:
      type: object
      required: [role]
      properties:
        role: { $ref: "#/components/schemas/Role" }
    UserList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/ManagedUser" }
...
    FieldError:
      type: object
      required: [field, code, message]
      properties:
        field: { type: string }
        code: { type: string }
        message: { type: string }
    Error:
      type: object
      required: [code, message, requestId]
      properties:
        code: { type: string }
        message: { type: string }
        requestId: { type: string }
        details:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"